[issue-223] Fixed even more tests!
